requirements.txt
arff
chardet
openml
pandas
PyYAML
Requests
pandas
numpy
kaggle
fasttext
autogluon
gensim
fasttext
tabpfn
seaborn
xgboost
skrub
shap